Implement Mehrotra Predictor-Corrector IPM for LPs.
Create an implementation of Mehrotra's Predictor Corrector Interior-Point Method for solving linear programs. This required the creation of a new class of linearized programs referred to as SlackLinearPrograms which are LPs in the standard form where there is a single equality constraint, Ax = b, and a positivity constraint on the state, x >= 0. The resulting algorithm produces an a modified state variable. In order to ensure that the output is uniform (at least for the state) we create conversion and reversion methods to switch between the internal and external states.