Update with g-entropic risk measures.
Create a more general class of risk measures on which to operate.
These risk measures, called g-entropic risk measures, are a
generalization of the conditional-value-at-risk (CVaR). We have
also implemented another example of this in the entropic-value-at-
risk (EVaR). The framework should allow for the creation of more
general risk functions.