Code cleanup and reorganization.
Clean up several unnecessarily public methods that were used to
generate simpler python bindings. These methods have now been set
to private to prevent any tampering.

Created several new routines to replace the task of extracting a
subset of rows from a sparse matrix in the Eigen library. This
task was performed repeatedly in the CRE algorithm and should
simplify the code base and make it easier to understand.

TODO: the implementation in CRE appends matrix pointers without
verifying whether or not the pointers are null. This needs to be
fixed before the updated code can be used.