Simplify the CRE subproblem code.
Creates two new files to augment Eigen's SparseMatrix and Matrix
operations, adding functions `get_rows`, `vstack`, and `vstack_rows` in
the util/DenseOps.h and util/SparseOps.h files. These functions are
templated to accept any Matrix (or SparseMatrix) type. The functionality
is as follows:

`get_rows` - extracts a subset of rows corresponding to true values in
a boolean array argument.
`vstack` - takes a vector of matrices as input and stacks them
vertically in the order in which they appear.
`vstack_rows` - acts as a combination of the above two methods and is
created to provide slightly higher performance than calling the above
individually. This method requires a vector of pairs corresponding to
the matrices to-be-stacked and their associated rows of interest.